+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- cosmwasm-vm: Add `Cache::save_wasm_unchecked` to save Wasm blobs that have
+  been checked before. This is useful for state-sync where we know the Wasm code
+  was checked when it was first uploaded. ([#1635])
+
+[#1635]: https://github.com/CosmWasm/cosmwasm/pull/1635
+
 ### Changed
 
 - cosmwasm-vm: Add checks for table section of Wasm blob ([#1631]).